Reducer moved to separated file
