+++ src/components/CategoriesTabs.tsx
@@ -2,44 +2,10 @@ import { useReducer } from 'react'
 import { Tabs } from '@mui/material'
 import CategoryBtn from './CategoryBtn'
 import { categoriesData } from '~/data/categories'
-import { createAction, ActionType, createReducer } from 'typesafe-actions'
-
-const setActiveCategory = createAction('SET_ACTIVE_CATEGORY')<{
-	activeCategory: number
-	activeSubCategory: number
-}>()
-
-const setActiveSubCategory = createAction('SET_ACTIVE_SUB_CATEGORY')<{
-	activeSubCategory: number
-}>()
-
-const actions = { setActiveCategory, setActiveSubCategory }
-
-type CategoriesAction = ActionType<typeof actions>
-
-interface State {
-	activeCategory: number
-	activeSubCategory: number
-}
-
-const initialState: State = {
-	activeCategory: 0,
-	activeSubCategory: 0,
-}
-
-const reducer = createReducer<State, CategoriesAction>(initialState)
-	.handleAction(setActiveCategory, (state, action) => ({
-		...state,
-		activeCategory: action.payload.activeCategory,
-		activeSubCategory: action.payload.activeSubCategory,
-	}))
-	.handleAction(setActiveSubCategory, (state, action) => ({
-		...state,
-		activeSubCategory: action.payload.activeSubCategory,
-	}))
+import { categoriesReducer, initialState, setActiveCategory, setActiveSubCategory } from '~/reducers/categoriesReducer'
 
 const CategoriesTabs = () => {
-	const [state, dispatch] = useReducer(reducer, initialState)
+	const [state, dispatch] = useReducer(categoriesReducer, initialState)
 
 	return (
 		<>
@@ -56,9 +22,7 @@ const CategoriesTabs = () => {
 						<CategoryBtn
 							key={category.id}
 							text={category.name}
-							onClick={() => {
-								dispatch(setActiveCategory({ activeCategory: category.id, activeSubCategory: 0 }))
-							}}
+							onClick={() => dispatch(setActiveCategory({ activeCategory: category.id, activeSubCategory: 0 }))}
 							isActive={state.activeCategory === category.id}
 						/>
 					))}
